services/finplanbot/requirements.txt
# requirements.txt — Python libraries your bot needs

python-telegram-bot
openai
pillow
